Move util methods to their own module
